be more lenient against input/iframe in head
There are trackers & third parties scripts out there that inject iframe
or input within head.

Because of that, we do not properly extract title & meta from a page
since they are technically not in <head> anymore.

But the user may not be able to fix this and it may be difficult to
explain why it works like this although the browsers does not care.